Move part of the vector tile sample code to a new Mapsui.Experimental.VectorTiles package
