+++ Mapsui.Experimental.VectorTiles/README.md
@@ -1,0 +1,4 @@
+﻿### Mapsui.Experimental.VectorTiles
+
+This is an experimental project to support vector tiles in Mapsui. It is not yet complete and may change 
+significantly in the future. This project is a layer on top of VexTile.TileSource.Mvt and BruTile.
