+++ Mapsui.slnx
@@ -316,6 +316,10 @@
     <BuildType Solution="Ad-Hoc|*" Project="Debug" />
     <BuildType Solution="AppStore|*" Project="Debug" />
   </Project>
+  <Project Path="Mapsui.Experimental.VectorTiles/Mapsui.Experimental.VectorTiles.csproj" Id="cbf48a30-76b7-44e7-a869-d9a5d7699230">
+    <BuildType Solution="Ad-Hoc|*" Project="Release" />
+    <BuildType Solution="AppStore|*" Project="Release" />
+  </Project>
   <Project Path="Mapsui.Extensions/Mapsui.Extensions.csproj">
     <BuildType Solution="Ad-Hoc|*" Project="Debug" />
     <BuildType Solution="AppStore|*" Project="Debug" />
